replace old hooks with db connection methods from ManagedConnectionBase
